+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 1.1.2 - 2024-03-27
+
+### Changed
+
+- Changed SourceMap type hints to SourceMapInterface.
+
+### Fixed
+
+- Fixed issue where router would check if a route file exists on a null directory.
+- Updated Pyncer HTTP function calls to use URI variants.
+
 ## 1.1.1 - 2023-09-05
 
 ### Changed